requirements.txt
streamlit
chromadb
sentence-transformers
google-generativeai
groq
python-dotenv
